Fix #4 boot crashes — wrong return types + JSON:API filter syntax
Three real bugs surfaced when installing the extension on a fresh
local Flarum 2 instance under strict MySQL 8 / php-fpm:

1. SocialGroupPostResource::deleting() and ::deleted() declared
   `: ?object` returning null. The parent (AbstractResource via
   HasHooks) declares both as `: void`. PHP strict signature
   checking refused to load the class, which prevented the resource
   from registering, which broke Flarum boot — every CLI command and
   HTTP request returned exit 255 / 500 with no output. Changed both
   hooks to `: void` to match the parent contract.

2. Every Resource scope() that I added read `?filter[group]=N` /
   `?filter[discussion]=N`. Flarum 2 AbstractDatabaseResource's
   `filters()` is final and throws — JSON:API filter[] syntax
   requires a registered Search\Filter class per param. Switched to
   plain query params (`?groupId=N`, `?discussionId=N`, `?q=foo`)
   that go straight through getQueryParams() in scope(); JS helpers
   updated to match.

3. The original social_group_post_reactions migration declared
   `user_id` as unsignedBigInteger. Flarum core's users.id is
   unsignedInteger (`increments()`) — strict MySQL 8 rejects the FK
   constraint with "Referencing column and referenced column are
   incompatible". Existing installs that ran the migration on
   pre-strict MySQL got the table with a BIGINT column and the FK
   silently coerced; fresh installs on MySQL 8 crash. Changed the
   column type to unsignedInteger so fresh installs work and the
   constraint type-matches. Existing installs are unaffected (the
   migration only runs once and is recorded in the migrations
   table).

Verified end-to-end on Herd Flarum 2 install:
- php flarum migrate         → DONE, exit 0
- php flarum cache:clear     → OK
- GET /api/social-groups                                 → 200
- GET /api/social-group-discussions?groupId=1            → 200
- GET /api/social-group-posts?discussionId=1             → 200
- GET /api/social-group-members?groupId=1 (guest)        → 401 (correct)
- GET /api/social-group-join-requests?groupId=1 (guest)  → 401 (correct)

+++ migrations/2024_01_01_000010_create_social_group_post_reactions_table.php
@@ -9,7 +9,12 @@ return [
             $schema->create('social_group_post_reactions', function (Blueprint $table) {
                 $table->bigIncrements('id');
                 $table->unsignedBigInteger('post_id');
-                $table->unsignedBigInteger('user_id');
+                // `users.id` is INT UNSIGNED in core (increments()), not
+                // BIGINT — declaring this as unsignedBigInteger made the
+                // FK constraint fail on strict MySQL 8 ("Referencing
+                // column and referenced column are incompatible"). Must
+                // match the referenced column type exactly.
+                $table->unsignedInteger('user_id');
                 $table->string('reaction', 20);
                 $table->timestamps();
